refactor/parsec: extract functional tests into a separate file
