+++ input_graphs/functional_tests_add_peer/alice.dot
