+++ input_graphs/functional_tests_add_peer/carol.dot
