+++ input_graphs/functional_tests_handle_malice_accomplice/alice.dot
