+++ input_graphs/functional_tests_handle_malice_duplicate_votes/carol.dot
