+++ input_graphs/functional_tests_handle_malice_genesis_event_creator_not_genesis_member/alice.dot
