+++ input_graphs/functional_tests_handle_malice_genesis_event_not_after_initial/alice.dot
