+++ input_graphs/functional_tests_handle_malice_handle_fork/bob.dot
