+++ input_graphs/functional_tests_handle_malice_invalid_accusation/carol.dot
