+++ input_graphs/functional_tests_handle_malice_invalid_gossip_creator/alice.dot
