+++ input_graphs/functional_tests_handle_malice_stale_other_parent/carol.dot
