+++ input_graphs/functional_tests_remove_peer/alice.dot
